introduce atomic block extension
