+++ plugin/evm/export_tx_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/ava-labs/coreth/utils"
 	"github.com/ava-labs/libevm/common"
 	"github.com/holiman/uint256"
+	"github.com/stretchr/testify/require"
 )
 
 // createExportTxOptions adds funds to shared memory, imports them, and returns a list of export transactions
@@ -457,7 +458,8 @@ func TestExportTxSemanticVerify(t *testing.T) {
 		}
 	}()
 
-	parent := vm.LastAcceptedBlockInternal().(*Block)
+	parent, ok := vm.LastAcceptedVMBlock().GetBlockExtension().(atomic.AtomicBlockContext)
+	require.True(t, ok)
 
 	key := testKeys[0]
 	addr := key.Address()
@@ -1708,7 +1710,7 @@ func TestNewExportTx(t *testing.T) {
 				}
 			}()
 
-			parent := tvm.vm.LastAcceptedBlockInternal().(*Block)
+			parent := tvm.vm.LastAcceptedVMBlock()
 			importAmount := uint64(50000000)
 			utxoID := avax.UTXOID{TxID: ids.GenerateTestID()}
 
@@ -1768,7 +1770,9 @@ func TestNewExportTx(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			parent = tvm.vm.LastAcceptedBlockInternal().(*Block)
+			parent = tvm.vm.LastAcceptedVMBlock()
+			atomicContextIntf, ok := parent.GetBlockExtension().(atomic.AtomicBlockContext)
+			require.True(t, ok)
 			exportAmount := uint64(5000000)
 
 			state, err := tvm.vm.blockChain.State()
@@ -1795,8 +1799,8 @@ func TestNewExportTx(t *testing.T) {
 			if err := exportTx.Visit(&atomic.SemanticVerifier{
 				Backend: backend,
 				Tx:      tx,
-				Parent:  parent,
-				BaseFee: parent.ethBlock.BaseFee(),
+				Parent:  atomicContextIntf,
+				BaseFee: parent.GetEthBlock().BaseFee(),
 			}); err != nil {
 				t.Fatal("newExportTx created an invalid transaction", err)
 			}
@@ -1877,7 +1881,7 @@ func TestNewExportTxMulticoin(t *testing.T) {
 				}
 			}()
 
-			parent := tvm.vm.LastAcceptedBlockInternal().(*Block)
+			parent := tvm.vm.LastAcceptedVMBlock()
 			importAmount := uint64(50000000)
 			utxoID := avax.UTXOID{TxID: ids.GenerateTestID()}
 
@@ -1967,7 +1971,9 @@ func TestNewExportTxMulticoin(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			parent = tvm.vm.LastAcceptedBlockInternal().(*Block)
+			parent = tvm.vm.LastAcceptedVMBlock()
+			atomicContextIntf, ok := parent.GetBlockExtension().(atomic.AtomicBlockContext)
+			require.True(t, ok)
 			exportAmount := uint64(5000000)
 
 			testKeys0Addr := testKeys[0].EthAddress()
@@ -1999,8 +2005,8 @@ func TestNewExportTxMulticoin(t *testing.T) {
 			if err := exportTx.Visit(&atomic.SemanticVerifier{
 				Backend: backend,
 				Tx:      tx,
-				Parent:  parent,
-				BaseFee: parent.ethBlock.BaseFee(),
+				Parent:  atomicContextIntf,
+				BaseFee: parent.GetEthBlock().BaseFee(),
 			}); err != nil {
 				t.Fatal("newExportTx created an invalid transaction", err)
 			}
